@@ -2,12 +2,13 @@
   "name": "nextjs-notion-starter-kit",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "description": "The perfect starter kit for building beautiful websites with Next.js and Notion.",
   "author": "Travis Fischer <travis@transitivebullsh.it>",
   "repository": "transitive-bullshit/nextjs-notion-starter-kit",
   "license": "MIT",
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
     "dev": "next dev",
@@ -28,6 +29,7 @@
     "@keyv/redis": "^2.2.3",
     "classnames": "^2.3.1",
     "date-fns": "^2.25.0",
+    "expiry-map": "^2.0.0",
     "fathom-client": "^3.0.0",
     "got": "^11.8.2",
     "isomorphic-unfetch": "^3.1.0",
@@ -35,16 +37,16 @@
     "lqip-modern": "^1.2.0",
     "next": "^12.1.0",
     "node-fetch": "^2.6.1",
-    "notion-client": "^5.0.0",
-    "notion-types": "^5.0.0",
-    "notion-utils": "^5.0.0",
-    "p-map": "^4.0.0",
-    "p-memoize": "^4.0.0",
+    "notion-client": "^6.0.0",
+    "notion-types": "^6.0.0",
+    "notion-utils": "^6.0.0",
+    "p-map": "^5.3.0",
+    "p-memoize": "^6.0.1",
     "react": "^17.0.2",
     "react-body-classname": "^1.3.1",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
-    "react-notion-x": "^5.0.0",
+    "react-notion-x": "^6.0.0",
     "react-static-tweets": "^0.7.1",
     "react-use": "^17.3.2",
     "static-tweets": "^0.7.1",
@@ -64,7 +66,6 @@
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.3.0",
     "next-compose-plugins": "^2.2.1",
-    "next-transpile-modules": "^9.0.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.4.1",
     "typescript": "^4.4.4"
